Add files via upload
The project is likely to involve developing a machine learning model to predict the price of cars based on various features such as brand, model, mileage, and engine type. The project may have used a dataset sourced from Kaggle, and the notebook could contain code written in Python using libraries like scikit-learn and pandas. Overall, the project aims to provide an accurate way of estimating the value of a car based on its characteristics.